// app/enums/enums.ts
export enum DRAWERROUTES {
  DASHBOARD = 'Explore',
  LOGIN = 'Login',
  REGISTRATION = 'Registration',
  NOTIFICATIONS = 'Notifications',
  SETTINGS = 'Settings',
  PAYMENT = 'Payment',
}

export enum AUTHROUTES {
  LOGINDEFAULT = 'LoginDefault',
  LOGINFORGOTPASSWORD = 'LoginForgotPassword',
  LOGINRESETPASSWORD = 'LoginResetPassword',
  REGISTER = 'Register',
}